Add files via upload
The main objective of the system is to build a placement Q&A session before the drive. The juniors can able to raise questions regarding the company and rounds. The seniors working in that company will post their answers. The faculty will approve/ reject the answers answered by the senior